Update Source/Picker.Date.js
I added three options to the PickerDate control. 
These affect the time picker specifically. 
The options are:
		minHour: 0,
		maxHour: 23,
		hourStart: '01',

If the user sets minHour and maxHour through a PickerDate object, 
he/she can set the time limits to choose times with timewheel, for example
officehours, 8AM to 6PM.

Then, the hourStart, a bit of a luxury, make it easier for the user to choose 
a time. Also, when a second PickerDate object, which is destined to specify an ending 
date and tie for an event, the time in the timewheel can be initialized to one hour after 
the PickarDate starting event object was selected to. 

Feedback?