Update Source/Picker.Date.Range.js
I added a new property for the Picker.Date.Range object: disableStart.




The idea behind this option is to prevent users from inadvertently changing the start date field value. Without this option, in cases where the Picker.Date.Range object has been initialized to start at a certain date, and if the user clicks more than once over the Picker.Date.Range control, the start date field value is likely to change. With the implementation of the disableStart option, and when this option is set to true, the start date input from the Picker.Date.Range control is attributed readonly, i.e. disables the user from selecting this input field, and despite the number of clicks over the Picker.Date.Range control, the start date field's value will remain the initialized value. 